+++ driver/src/locks/xaie_locks_aieml.c
@@ -154,7 +154,7 @@ AieRC _XAieMl_LockSetValue(XAie_DevInst *DevInst, const XAie_LockMod *LockMod,
 	u32 RegVal;
 
 	RegAddr = (u64)(LockMod->LockSetValBase +
-		LockMod->LockSetValOff * Lock.LockId) +
+		LockMod->LockSetValOff * (u64)Lock.LockId) +
 		XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col);
 
 	RegVal = XAie_SetField(Lock.LockVal, LockMod->LockInit->Lsb,
